+++ Bin/NodeGraphs/Audio/AudioNodeGraph.json
@@ -1,1 +1,1 @@
-{"Connections":[{"NodeFrom":5,"NodeTo":4,"SlotFrom":0,"SlotTo":0},{"NodeFrom":4,"NodeTo":0,"SlotFrom":0,"SlotTo":0}],"InlineInputValues":[{"NodeId":5,"SlotId":0,"Value":"Gameplay"}],"Nodes":[{"Id":0,"IsImmortal":true,"NodeType":"Final Mix","PositionX":1000.0,"PositionY":400.0},{"Id":3,"IsImmortal":false,"NodeType":"Echo Filter","PositionX":213.0,"PositionY":488.0},{"Id":4,"IsImmortal":false,"NodeType":"Mix","PositionX":690.0,"PositionY":333.0},{"Id":5,"IsImmortal":false,"NodeType":"Source Bus","PositionX":110.0,"PositionY":276.0}]}
+{"Connections":[{"NodeFrom":2,"NodeTo":4,"SlotFrom":0,"SlotTo":0},{"NodeFrom":4,"NodeTo":0,"SlotFrom":0,"SlotTo":0}],"InlineInputValues":[{"NodeId":2,"SlotId":0,"Value":"Gameplay"}],"Nodes":[{"Id":0,"NodeType":"Final Mix","PositionX":1000.0,"PositionY":300.0},{"Id":4,"NodeType":"Mix","PositionX":482.0,"PositionY":248.0},{"Id":2,"NodeType":"Source Bus","PositionX":63.0,"PositionY":301.0}]}
