+++ WonderMake/Audio/AudioMixingNodeGraph.h
@@ -21,12 +21,9 @@ public:
 protected:
 	virtual void RegisterNodes() override;
 
-	virtual void FirstTimeSetup() override;
-
 	void PostLoad() override;
 
 private:
-	SNode* myRootNode;
 
 	std::vector<SoLoud::Bus> myBusses;
