+++ WonderMake/Logging/LogMessageTags.h
@@ -3,4 +3,6 @@
 
 static const std::string TagError = "[Error]";
 static const std::string TagWarning = "[Warning]";
+static const std::string TagSuccess = "[Success]";
 static const std::string TagOpenGL = "[OpenGL]";
+static const std::string TagNodeGraph = "[NodeCompilation]";
