+++ WonderMake/NodeGraph/NodeGraph.cpp
@@ -1,5 +1,6 @@
 #include "pch.h"
 #include "NodeGraph.h"
+#include "Debugging/DebugSettingsSystem.h"
 
 NodeGraph::NodeGraph(std::filesystem::path aFilePath)
 	: myPath(aFilePath)
@@ -9,6 +10,18 @@ NodeGraph::NodeGraph(std::filesystem::path aFilePath)
 	uniqueIdCounter++;
 }
 
+void NodeGraph::CompileExternal()
+{
+	Compile();
+
+	SystemPtr<DebugSettingsSystem> debugSettings;
+
+	const bool shouldSave = debugSettings->GetOrCreateDebugValue("NodeGraph/Should save on compile", true);
+
+	if (shouldSave)
+		Save();
+}
+
 void NodeGraph::Save()
 {
 	nlohmann::json json = Serialize();
@@ -16,11 +29,28 @@ void NodeGraph::Save()
 	std::ofstream file(myPath);
 
 	file << std::setw(4) << json.dump();
+
+	WmLog(TagSuccess, TagNodeGraph, "Saved node graph [", myPath.string(), "]");
+}
+
+void NodeGraph::ExecuteExternal()
+{
+	SystemPtr<DebugSettingsSystem> debugSettings;
+
+	const bool shouldCompile = debugSettings->GetOrCreateDebugValue("NodeGraph/Should compile before execute", true);
+
+	if (shouldCompile)
+		Compile();
+
+	Execute();
 }
 
 void NodeGraph::Load()
 {
 	RegisterNodes();
+
+	if (myRootNodeType == nullptr)
+		WmLog(TagError, TagNodeGraph, "No root node type registered! Things may break.");
 
 	std::ifstream file(myPath, std::fstream::app);
 
@@ -40,6 +70,66 @@ void NodeGraph::Load()
 	}
 
 	PostLoad();
+}
+
+void NodeGraph::KillNodesThatWantToDie()
+{
+	plf::colony<SNode>::colony_iterator<false> it = Nodes.begin();
+
+	while (it != Nodes.end())
+	{
+		//[Nicos]: Root nodes can't die even if they want to
+		if (it->IWantToDie && !IsNodeTypeRoot(it->NodeType))
+			it = KillNode(it);
+		else
+			it++;
+	}
+}
+
+bool NodeGraph::IsNodeTypeRoot(const SNodeTypeBase& aNodeType) const
+{
+	return &aNodeType == myRootNodeType;
+}
+
+plf::colony<SNode>::colony_iterator<false> NodeGraph::KillNode(plf::colony<SNode>::colony_iterator<false> aIterator)
+{
+	for (auto& inputSlot : aIterator->InputSlotInstances)
+	{
+		if (inputSlot->Connection)
+		{
+			auto it = Connections.get_iterator_from_pointer(inputSlot->Connection);
+			if (it != Connections.end())
+				Connections.erase(it);
+		}
+	}
+
+	for (auto& outputSlot : aIterator->OutputSlotInstances)
+	{
+		for (auto&& connection : outputSlot->Connections)
+		{
+			if (connection)
+			{
+				auto it = Connections.get_iterator_from_pointer(connection);
+				if (it != Connections.end())
+					Connections.erase(it);
+			}
+		}
+	}
+
+	return Nodes.erase(aIterator);
+}
+
+void NodeGraph::FirstTimeSetup()
+{
+	SetupRootNode();
+}
+
+void NodeGraph::SetupRootNode()
+{
+	if (!myRootNodeType)
+		return;
+
+	myRootNode = AddNode(myRootNodeType->Title, { 1000.f, 300.f });
 }
 
 nlohmann::json NodeGraph::Serialize()
@@ -92,11 +182,29 @@ void NodeGraph::Deserialize(const nlohmann::json& aJsonFile)
 	{
 		DeserializeInput(it.value());
 	}
+
+	if (!myRootNode && myRootNodeType)
+	{
+		for (auto& node : Nodes)
+		{
+			if (&node.NodeType == myRootNodeType)
+			{
+				myRootNode = &node;
+				break;
+			}
+		}
+
+		if (!myRootNode)
+		{
+			WmLog(TagError, TagNodeGraph, "No root node detected on load! Adding one...");
+			SetupRootNode();
+		}
+	}
 }
 
-void NodeGraph::SerializeNode(SNode& aNode, nlohmann::json& aJson)
+void NodeGraph::SerializeNode(const SNode& aNode, nlohmann::json& aJson)
 {
-	aJson.push_back(aJson.object({ {"NodeType", aNode.NodeType.Title}, {"IsImmortal", aNode.IsImmortal}, {"PositionX", aNode.Position.x}, {"PositionY", aNode.Position.y}, {"Id", aNode.Id} }));
+	aJson.push_back(aJson.object({ {"NodeType", aNode.NodeType.Title}, {"PositionX", aNode.Position.x}, {"PositionY", aNode.Position.y}, {"Id", aNode.Id} }));
 }
 
 void NodeGraph::DeserializeNode(const nlohmann::json& aJson)
@@ -105,13 +213,13 @@ void NodeGraph::DeserializeNode(const nlohmann::json& aJson)
 
 	if (node)
 	{
-		node->IsImmortal = aJson["IsImmortal"].get<bool>();
 		node->Position.x = aJson["PositionX"].get<f32>();
 		node->Position.y = aJson["PositionY"].get<f32>();
 		node->Id = aJson["Id"].get<size_t>();
 
-		if (node->Id >= NextNodeIndex)
-			NextNodeIndex = node->Id + 1;
+		//keep node index updated so that we don't accidentally add a new node with an existing id
+		if (node->Id >= NextNodeId)
+			NextNodeId = node->Id + 1;
 	}
 }
 
@@ -151,13 +259,13 @@ void NodeGraph::DeserializeConnection(const nlohmann::json& aJson)
 	connectionRef.InputSlot->Connection = &connectionRef;
 }
 
-SNode* NodeGraph::AddNode(std::string NodeType)
+SNode* NodeGraph::AddNode(const std::string aNodeType, const ImVec2 aLocation)
 {
 	for (auto& registeredNode : RegisteredNodes)
 	{
-		if (registeredNode.Name == NodeType)
+		if (registeredNode.Name == aNodeType)
 		{
-			return &registeredNode.AddNodeLambda({});
+			return &registeredNode.AddNodeLambda(aLocation);
 		}
 	}
 
@@ -175,7 +283,7 @@ SNode* NodeGraph::FindNodeById(const size_t aId)
 	return nullptr;
 }
 
-void NodeGraph::CompileNodeGraph(SNode& aRoot, std::vector<SCompiledNode>& aNodeStack)
+void NodeGraph::CompileNodeGraph(SNode& aRoot, std::vector<SCompiledNode>& aNodeStack, const bool aIsFirstCompileCall)
 {
 	aNodeStack.push_back({ aRoot });
 
@@ -187,11 +295,12 @@ void NodeGraph::CompileNodeGraph(SNode& aRoot, std::vector<SCompiledNode>& aNode
 			SNode* node = static_cast<SNode*>(id);
 
 			if (node)
-			{
-				CompileNodeGraph(*node, aNodeStack);
-			}
+				CompileNodeGraph(*node, aNodeStack, false);
 		}
 	}
+
+	if (aIsFirstCompileCall)
+		WmLog(TagSuccess, TagNodeGraph, "Compiled node graph [", myPath.string(), "]!");
 }
 
 void NodeGraph::SerializeInlineInputs(SNode& aNode, json& aInputArray)
@@ -203,7 +312,7 @@ void NodeGraph::SerializeInlineInputs(SNode& aNode, json& aInputArray)
 		if (inputSlotInstance->HasConnection())
 			return;
 
-		inputSlotInstance->SerializeInlineInput(aNode.Id, i, aInputArray);
+		inputSlotInstance->SerializeInlineInput(static_cast<i32>(aNode.Id), i, aInputArray);
 	}
 }
